EPMRDRUNET-1781 MVC Fixed paging, searchLine and added lite authentification and lite authorisation

+++ Epam.Library.Dal.Contracts/IRoleDao.cs
@@ -7,7 +7,7 @@ namespace Epam.Library.Dal.Contracts
     {
         IEnumerable<Role> GetAll();
 
-        Role GetById(long id);
+        Role GetById(int id);
 
         Role GetByName(string name);
     }
